Refactor task components
