+++ src/routes/+page.svelte
@@ -1,11 +1,9 @@
 <script lang="ts">
 	import Score from '$lib/components/Score.svelte';
-	import TaskList from '$lib/components/TaskList.svelte';
-	import WritingTask from '$lib/components/WritingTask.svelte';
+	import TaskList from '$lib/components/tasks/TaskList.svelte';
 	import { getWordIds, type TaskName } from '$lib/api';
 	import { onMount } from 'svelte';
-	import ListeningTask from '$lib/components/ListeningTask.svelte';
-	import SpeechTask from '$lib/components/SpeechTask.svelte';
+	import TaskHandler from '$lib/components/tasks/TaskHandler.svelte';
 
 	let score = 0;
 	let maxScore = 0;
@@ -41,18 +39,10 @@
 </script>
 
 <TaskList bind:tasks />
-
-<!--It ain't pretty, could probaby be refactored.-->
 {#if tasks.length > 0 && currentWordId}
 	{#key currentTask}
 		{#key currentWordId}
-			{#if currentTask === 'listening'}
-				<ListeningTask on:check={onCheck} wordId={currentWordId} />
-			{:else if currentTask === 'speech'}
-				<SpeechTask on:check={onCheck} wordId={currentWordId} />
-			{:else}
-				<WritingTask on:check={onCheck} task={currentTask} wordId={currentWordId} />
-			{/if}
+			<TaskHandler bind:task={currentTask} bind:wordId={currentWordId} on:check={onCheck} />
 		{/key}
 	{/key}
 {:else}
